+++ CHANGES.md
@@ -1,6 +1,10 @@
 # CHANGES
 ## RosarioSIS Student Information System
 
+Changes in 11.5.3
+-----------------
+- Fix MySQL 5.7.5+ error due to ONLY_FULL_GROUP_BY mode: use MIN() in Address.inc.php
+
 Changes in 11.5.2
 -----------------
 - Fix get Student Info when Parent associated to various students in ReportCards.fnc.php